requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
pydantic==2.12.3
pytesseract==0.3.13
sentence-transformers==5.1.2
python-multipart==0.0.20
pyzbar==0.1.9
langchain-experimental==0.3.4
faiss-cpu==1.12.0